add workflow types recognition
